requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pytesseract
Pillow
pdfplumber
openpyxl
reportlab